+++ app/src/main/java/woowacourse/movie/movie/Movie.kt
@@ -1,6 +1,7 @@
-package woowacourse.movie
+package woowacourse.movie.movie
 
 import androidx.annotation.DrawableRes
+import woowacourse.movie.R
 import java.io.Serializable
 import java.time.LocalDate
