+++ app/src/main/java/woowacourse/movie/movie/MovieBookingSeatInfo.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.movie
 
 import java.io.Serializable
