+++ app/src/main/java/woowacourse/movie/movie/MovieMockData.kt
@@ -1,5 +1,6 @@
-package woowacourse.movie
+package woowacourse.movie.movie
 
+import woowacourse.movie.R
 import java.time.LocalDate
 
 object MovieMockData {
