+++ app/src/main/java/woowacourse/movie/movielist/listener/AdOnClickListener.kt
@@ -1,8 +1,9 @@
-package woowacourse.movie
+package woowacourse.movie.movielist.listener
 
 import android.content.Context
 import android.content.Intent
 import android.net.Uri
+import woowacourse.movie.Ad
 
 class AdOnClickListener(private val context: Context) : OnClickListener<Ad> {
     override fun onClick(item: Ad) {
