+++ app/src/main/java/woowacourse/movie/movielist/listener/MovieOnClickListener.kt
@@ -1,7 +1,9 @@
-package woowacourse.movie
+package woowacourse.movie.movielist.listener
 
 import android.content.Context
+import woowacourse.movie.BundleKeys
 import woowacourse.movie.activity.MovieDetailActivity
+import woowacourse.movie.movie.Movie
 
 class MovieOnClickListener(private val context: Context) : OnClickListener<Movie> {
     override fun onClick(item: Movie) {
